+++ inst/shiny/server.R
@@ -323,8 +323,8 @@ shiny::shinyServer(function(input, output, session) {
                          "PHASE IDs" = as.list(paste0(filedata3()[[3]][[2]], ": ", filedata3()[[3]][[1]])))
 
       #Order alphabetically
-      refs_force[[1]] <- refs_force[[1]][order(unlist(refs_force[[1]]))]
-      refs_force[[2]] <- refs_force[[2]][order(unlist(refs_force[[2]]))]
+      #refs_force[[1]] <- refs_force[[1]][order(unlist(refs_force[[1]]))]
+      #refs_force[[2]] <- refs_force[[2]][order(unlist(refs_force[[2]]))]
 
       return(shiny::selectInput("selectFORCE",
                                 label = NULL,
@@ -340,8 +340,8 @@ shiny::shinyServer(function(input, output, session) {
         refs_choices <- list("PHASE_NAMES" = as.list(paste0(filedata3()[[3]][[2]])),
                              "PHASE IDs" = as.list(paste0(filedata3()[[3]][[2]], ": ", filedata3()[[3]][[1]])))
         #Order alphabetically
-        refs_choices[[1]] <- refs_choices[[1]][order(unlist(refs_choices[[1]]))]
-        refs_choices[[2]] <- refs_choices[[2]][order(unlist(refs_choices[[2]]))]
+        #refs_choices[[1]] <- refs_choices[[1]][order(unlist(refs_choices[[1]]))]
+        #refs_choices[[2]] <- refs_choices[[2]][order(unlist(refs_choices[[2]]))]
 
         return(shiny::updateSelectInput(session, "selectPHASES_fps",
                                         choices = refs_choices,
@@ -461,6 +461,8 @@ shiny::shinyServer(function(input, output, session) {
 
       fps_reactive <- shiny::eventReactive(input$goButton_fps, {
 
+        shiny::withProgress(message = "Computing...", value = 1,{
+
         if (input$std_conc_check_fps == FALSE) {
           std_conc_fps <- NA
         }
@@ -507,6 +509,8 @@ shiny::shinyServer(function(input, output, session) {
                              amorphous = sub(".*: ", "", input$selectAMORPH))
           return(afps_out)
         }
+
+        }) #end with progress
       })
 
       observe({
@@ -714,6 +718,8 @@ shiny::shinyServer(function(input, output, session) {
     shiny::observe({
       fps_reactive_editor <- shiny::eventReactive(input$goButton_editor, {
 
+        shiny::withProgress(message = "Computing...", value = 1,{
+
         if(input$std_conc_check_editor == FALSE) {
           std_conc_editor <- NA
           } else {
@@ -748,6 +754,9 @@ shiny::shinyServer(function(input, output, session) {
                               shift = input$shift_editor,
                               obj = input$selectOBJ_editor,
                               solver = input$selectSolver_editor)
+
+        }) #end withProgress
+
         })
 
       shiny::observe({